This commit adds the card screen and the option to choose between the card screen and the map screen.
