+++ package.json
@@ -57,6 +57,7 @@
     "react-leaflet": "^1.2.2",
     "react-resize-detector": "^0.5.0",
     "react-router": "^3.0.2",
+    "react-theme-provider": "^0.1.3",
     "sass-loader": "^6.0.5",
     "style-loader": "^0.18.2"
   }
